% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive6class.R
\name{Reactive6Class}
\alias{Reactive6Class}
\title{Defines a reactive class.}
\usage{
Reactive6Class(
  classname = NULL,
  public = list(),
  private = NULL,
  active = NULL,
  reactive = list(),
  inherit = NULL,
  parent_env = parent.frame(),
  ...
)
}
\description{
It is like a R6 class, but allows to specify a new field "reactive"
as a list of fields associated to an initialisation value.
The fields will be exported in the public interface and will behave
as reactive fields in a reactiveValues object.
}
\examples{
AppModel <- Reactive6Class(
  "AppModel",
  public = list(
    increment = function(value) {
      self$amount <- self$amount + value
    },
    asSquared = function() {
      return(self$amount * self$amount)
    },
    nonreactive_amount = 10
  ),
  reactive = list(
    amount = 10
  ),
)

In the above example
}
